backend/generate-email/requirements.txt
openai==1.54.0
httpx==0.27.0
psycopg2-binary==2.9.9
